% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bn_graph}
\alias{bn_graph}
\title{Create a DiagrammeR graph from fitted Bayesian Network}
\usage{
bn_graph(x)
}
\arguments{
\item{x}{a \code{bn.fit} object}
}
\value{
a graph object of class \code{dgr_graph}
}
\description{
Creates a \code{dgr_graph} object (cf. \link[DiagrammeR:create_graph]{DiagrammeR::create_graph})
from a \code{bn.fit} object (cf. \link[bnlearn:bn.fit]{bnlearn::bn.fit}).
}
\examples{
data("pavement_model", package = "shinyBN")
pavement_graph <- bn_graph(pavement_model)
# DiagrammeR::render_graph(pavement_graph)
}
