% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyBN.R
\name{launch_shinyBN}
\alias{launch_shinyBN}
\title{Launch the 'shinyBN' app}
\usage{
launch_shinyBN(object, app_title = NULL, node_labels = NULL,
  value_labels = NULL, help_texts = NULL, documentation_md = NULL,
  ...)
}
\arguments{
\item{object}{a discrete Bayesian network}

\item{app_title}{name to use for the instance of the app}

\item{node_labels}{optional named vector of labels to use for the nodes}

\item{value_labels}{optional named list of labels to use for the possible
values of the nodes}

\item{help_texts}{optional named character vector of helpful text to
include in the node interfaces}

\item{documentation_md}{optional path to a Markdown document to include in
the app; if provided, it is wrapped in \code{withMathJax}, so the documentation
can include math}

\item{...}{parameters to pass to \link[shiny:shinyApp]{shiny::shinyApp}}
}
\description{
Launch the Shiny-powered app to visualize and interact with a
\code{bn.fit} object (cf. \link[bnlearn:bn.fit]{bnlearn::bn.fit}).
}
\examples{
\dontrun{
launch_shinyBN(pavement_model)

doc <- system.file("extdata", "pavement_example.md", package = "shinyBN")
title <- "Interactive 'Pavement' Bayesian network with ShinyBN"
launch_shinyBN(pavement_model, app_title = title, documentation_md = doc)
}
}
