% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyBN.R
\docType{package}
\name{shinyBN-package}
\alias{shinyBN-package}
\alias{shinyBN}
\title{shinyBN: Interactive Visualization for Bayesian Networks}
\description{
Shiny-powered tool for visualizing probabilities in
Bayesian Networks.
}
