% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\docType{data}
\name{tasting_model}
\alias{tasting_model}
\title{Taste tests example model}
\format{An object of class \code{bn.fit} (inherits from \code{bn.fit.dnet}) of length 7.}
\usage{
tasting_model
}
\description{
Example of a Bayesian network with 7 nodes: 1 for underlying
product quality, 3 for taste testers' accuracy, and 3 for taste testers'
judgements. Explore the model with \code{launch_shinyBN_demo("tasting")}
}
\keyword{datasets}
